fixed bug in resolution-scale-dependancy

+++ Raytracer/Object.pde
@@ -16,11 +16,11 @@ class Object {
 }
 
 class Sphere extends Object { 
-  int radius;
+  float radius;
 
   Sphere(PVector origin_, int radius_, Material material_) {
     origin = PVector.mult(origin_, u);
-    radius = radius_ * int(u);
+    radius = radius_ * (u);
     c = material_.c;
     roughness = material_.roughness;
     reflectivity = material_.reflectivity;
